Brings tracr_l3 to within 5 commits of simpler main (it was 63 behind) and
picks up 39ce891d, the commit pypto main pins, which unblocks the pypto
submodule bump.

Seven conflicts. Two were more than a side-pick:

CallConfig gained an EIGHTH int32. Both branches had independently added a
seventh: tracr_l3 `flow_id` (TraCR host->device arrows), tracr
`capture_clock_anchors`. Keeping both invalidates three things a side-pick
leaves broken, because the mailbox config is a raw memcpy of the struct:
  - the static_assert pinning the wire layout still read 7 * sizeof(int32_t)
    (it caught the drift, which is why it exists) -> now 8;
  - worker.py's _CFG_FMT still declared 7 int32 slots, so every field after
    them would have been misread -> now 8;
  - the unpack tuple carries both fields, in declaration order.
Verified against the compiler rather than reasoned about: flow_id at 24,
capture_clock_anchors at 28, runtime_env at 32, sizeof 1152, and Python's
_CFG_FMT.size agrees at every one.

orchestrator.py: main changed the provenance-check signature and made
member_checks a 2-tuple. Took main's structure and re-wove the TraCR markers
and flow-id stamping around it; HEAD's 3-tuple unpack and device_args
argument no longer exist.

tracr removed X(Barrier) with INDEP_ORCH and never had CopyH2D / CopyD2H, so
our side is kept: Barrier 17, CopyH2D 18, CopyD2H 19 -- the ids the D3 AICore
lane hard-codes.

Verified: all four platforms build with BUILD_TRACR=ON; simpler.worker and
simpler.orchestrator import; both config fields round-trip through the
bindings.

+++ src/common/host_build_graph/graph_cache.h
@@ -16,7 +16,7 @@
 
 #include <type_traits>
 
-#include "task_id.h"
+#include "host_build_graph/task_id.h"
 #include "host_build_graph/types.h"
 
 struct GraphScopeResult {
@@ -66,8 +66,7 @@ constexpr uint64_t graph_const_hash_impl(const char *s, uint64_t h) {
 constexpr uint64_t GRAPH_KEY(const char *s) { return graph_const_hash_impl(s, 1469598103934665603ULL); }
 
 inline bool rt_graph_args_cacheable(const GraphTaskArgs &args) {
-    if (args.has_error || args.tensor_count() <= 0 ||
-        args.tensor_count() > static_cast<int32_t>(GRAPH_MAX_TENSOR_ARGS)) {
+    if (args.has_error || args.tensor_count() <= 0 || args.tensor_count() > GRAPH_MAX_TENSOR_ARGS) {
         return false;
     }
     for (int32_t i = 0; i < args.tensor_count(); ++i) {
